break-up code, remove comments, minor changes
