+++ clients/rust/examples/sysadmin-example.rs
@@ -8,6 +8,7 @@ use std::time::Duration;
 // see https://docs.rs/error-chain
 use sysadmin_bindings::error_chain_generated_errors::*;
 use sysadmin_bindings::SysAdminClient;
+
 // Available Commands:
 //     Set
 //     Get
@@ -44,12 +45,13 @@ fn main() {
 // Note the "Result" being returned is a type alias generated by error-chaing
 // https://docs.rs/error-chain/0.11.0/error_chain/example_generated/type.Result.html
 fn run() -> Result<(())> {
-
     let mut client = SysAdminClient::new(Duration::from_secs(2_u64), 1_u32, 1_u32);
 
     // init connection
     // if the connection fails we will report an error and exit
-    // client.connect("127.0.0.1:4000")?;
+    client.connect("127.0.0.1:4000").chain_err(|| {
+        "this will fail if sysadmin isn't running locally".to_string()
+    })?;
     let set_resp = client.set(String::from("foo"), 55_i32)?;
     println!("{:?}", set_resp);
 
@@ -58,19 +60,14 @@ fn run() -> Result<(())> {
     println!("{:?}", com_resp);
 
     let get_resp = client.get(String::from("foo"))?;
-
-    match get_resp {
-        sysadmin_bindings::GetResponse { id, status, kvs } => {
-            println!("GetResponse.id {:?}", id);
-            println!("GetResponse.status {:?}", status);
-            kvs.into_iter().for_each(|k| {
-                println!(
-                    "GetResponse.kvs key={:?} value={:?}",
-                    k.key,
-                    k.value.unwrap()
-                )
-            });
-        }
-    }
+    println!("GetResponse.id {:?}", get_resp.id);
+    println!("GetResponse.status {:?}", get_resp.status);
+    get_resp.kvs.into_iter().for_each(|k| {
+        println!(
+            "GetResponse.kvs key={:?} value={:?}",
+            k.key,
+            k.value.unwrap()
+        )
+    });
     Ok(())
 }
